feat: add event and project schemas with relationships to users and categories
